# Reduce allocs in ReadUint64 by pre-allocating byte buffer