PR captures the reasoning behind this:
https://github.com/containerd/cgroups/pull/275/files#r1152562543


    Reduce allocs in ReadUint64 by pre-allocating byte buffer
    
    Before
    
    Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^BenchmarkReaduint64$ github.com/containerd/cgroups/v3/cgroup1 -count=1 -timeout 30m
    
    goos: linux
    goarch: amd64
    pkg: github.com/containerd/cgroups/v3/cgroup1
    cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
    BenchmarkReaduint64-8             105118              9762 ns/op             848 B/op          5 allocs/op
    PASS
    ok      github.com/containerd/cgroups/v3/cgroup1        1.151s
    
    After
    
    Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^BenchmarkReaduint64$ github.com/containerd/cgroups/v3/cgroup1 -count=1 -timeout 30m
    
    goos: linux
    goarch: amd64
    pkg: github.com/containerd/cgroups/v3/cgroup1
    cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
    BenchmarkReaduint64-8             141484              8322 ns/op             128 B/op          3 allocs/op
    PASS
    ok      github.com/containerd/cgroups/v3/cgroup1        1.274s