@kolyshkin 2023-04-04
> Before
> 
> goos: linux goarch: amd64 pkg: github.com/containerd/cgroups/v3/cgroup1 cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz BenchmarkPids-8 59758 19905 ns/op 2064 B/op 15 allocs/op

Not sure why you use `BenchmarkPids` before and `BenchmarkReaduint64` after.