@kolyshkin 2023-04-04
I see number of allocations going down from 5 to 3, when using Go 1.20.2 (or 6 to 4 when using Go 1.19.6). This is far from savings reported in the PR description (which is obviously wrong as it says `BenchmarkPid` for "before" measurement).

Out of these 2 allocations, one comes from the `f.Stat` call from `os.OpenFile`. Indeed, we don't need `stat()` when working with cgroup files. The other, I guess, comes from inlining (if I move file reading to a separate function, the number of allocs goes up from 3 to 4).

So yes, the commit makes sense, the description though needs to be modified to
 - show the correct before and after measurements
 - tell that we avoid os.OpenFile because:
   - it calls os.Stat which we do not need
   - it pre-allocates 512 bytes while we're fine with 80

In addition, I would change 80 to 128 for alignment.