@kolyshkin 2023-04-04
Finally, note that there are a few other implementations in this repo, that are very similar to `readUint`. They all are in cgroup2 package, I was able to find two:
 - `getStatFileContentUint64` (uses `os.ReadFile`)
 - `readSingleFile` (uses `io.ReadAll`)
 
 I think those needs to be either consolidated, or fixed in a similar manner. Otherwise, we'll have the same problem in a couple of years once everyone has switched to cgroup v2.