@dcantah 2023-04-04
@kolyshkin Oop, you beat me to your last comment. I was drafting an issue for similar fixes we could make here, have the same two you found so far, may be more.