@dcantah 2023-04-04
> Technically speaking, the benchmark should use a real cgroupfs file (or any other file from /proc). The reason is, normal files have size, and os.ReadFile uses that information to prepare the buffer of the size needed. OTOH many cgroupfs (and /proc) files report the size of 0, which makes os.ReadFile to allocate 512 bytes buffer.

Probably is better to test on some magic file so +1, but given the change here is for us to provide the buffer, the `ReadFile` behavior doesn't come into the picture right?