@kolyshkin 2023-04-05
> Probably is better to test on some magic file so +1, but given the change here is for us to provide the buffer, the `ReadFile` behavior doesn't come into the picture right?

The `ReadFile` is used in the "before" benchmark.