@brancz 2023-04-06
Ah awesome, I was going to get back to this today, glad you already worked it out! Re-running the benchmarks I now see a total improvement of ~25%!

```
$ benchstat old.txt new.txt
name          old time/op    new time/op    delta
Readuint64-8    5.19µs ± 2%    3.87µs ± 1%  -25.48%  (p=0.000 n=10+10)

name          old alloc/op   new alloc/op   delta
Readuint64-8      840B ± 0%      120B ± 0%  -85.71%  (p=0.000 n=10+10)

name          old allocs/op  new allocs/op  delta
Readuint64-8      5.00 ± 0%      3.00 ± 0%  -40.00%  (p=0.000 n=10+10)
```

That means we should reap a total improvement of about 4%! Awesome!

I'm going to go ahead and close my PR.